prof: memoize TableRow render with dumb component
